[DOP-8140] Split Greenplum, Greenplum.Dialect and .*Options to separated submodules
