+++ docs/connection/db_connection/greenplum/connection.rst
@@ -1,0 +1,9 @@
+.. _greenplum-connection:
+
+Greenplum connection
+====================
+
+.. currentmodule:: onetl.connection.db_connection.greenplum.connection
+
+.. autoclass:: Greenplum
+    :members: get_packages, check
