+++ docs/connection/db_connection/greenplum/execute.rst
@@ -1,0 +1,17 @@
+.. _greenplum-execute:
+
+Executing statements in Greenplum
+==================================
+
+.. currentmodule:: onetl.connection.db_connection.greenplum.connection
+
+.. automethod:: Greenplum.fetch
+.. automethod:: Greenplum.execute
+.. automethod:: Greenplum.close
+
+.. currentmodule:: onetl.connection.db_connection.jdbc_mixin.options
+
+.. autopydantic_model:: JDBCOptions
+    :member-order: bysource
+    :model-show-field-summary: false
+    :field-show-constraints: false
