+++ docs/connection/db_connection/greenplum/index.rst
@@ -1,11 +1,19 @@
 .. _greenplum:
 
-Greenplum connector
+Greenplum
 ====================
 
 .. toctree::
     :maxdepth: 1
-    :caption: Greenplum connector
+    :caption: Connection
 
     prerequisites
-    greenplum
+    connection
+
+.. toctree::
+    :maxdepth: 1
+    :caption: Options
+
+    read
+    write
+    execute
