+++ docs/connection/db_connection/greenplum/write.rst
@@ -1,0 +1,13 @@
+.. _greenplum-write:
+
+Writing to Greenplum
+=====================
+
+For writing data to Greenplum, use :obj:`DBWriter <onetl.db.db_writer.db_writer.DBWriter>` with options below.
+
+.. currentmodule:: onetl.connection.db_connection.greenplum.options
+
+.. autopydantic_model:: GreenplumWriteOptions
+    :member-order: bysource
+    :model-show-field-summary: false
+    :field-show-constraints: false
