+++ setup.cfg
@@ -314,9 +314,6 @@ per-file-ignores =
         WPS219,
 # WPS437: Found protected attribute usage: spark._jvm
         WPS437,
-    onetl/connection/db_connection/greenplum.py:
-# WPS437 Found protected attribute usage: self.Dialect._
-        WPS437,
     onetl/connection/db_connection/kafka/connection.py:
 # WPS342: Found implicit raw string \\n
         WPS342,
