+++ tests/tests_unit/tests_db_connection_unit/test_db_options_unit.py
@@ -41,8 +41,8 @@ def test_db_options_connection_parameters_cannot_be_passed(options_class, arg, v
         (Postgres.ReadOptions, "ReadOptions", {"fetchsize": 10, "keytab": "a/b/c"}),
         (Postgres.WriteOptions, "JDBCWriteOptions", {"if_exists": "replace_entire_table", "keytab": "a/b/c"}),
         (Postgres.Options, "Options", {"if_exists": "replace_entire_table", "keytab": "a/b/c"}),
-        (Greenplum.ReadOptions, "ReadOptions", {"partitions": 10}),
-        (Greenplum.WriteOptions, "WriteOptions", {"if_exists": "replace_entire_table"}),
+        (Greenplum.ReadOptions, "GreenplumReadOptions", {"partitions": 10}),
+        (Greenplum.WriteOptions, "GreenplumWriteOptions", {"if_exists": "replace_entire_table"}),
     ],
 )
 def test_db_options_warn_for_unknown(options_class, options_class_name, known_options, caplog):
